+++ src/ui/ShortcutPathConverterDialog.cpp
@@ -829,8 +829,7 @@ bool ShortcutPathConverterDialog::RunModal() {
 
     window_presentation::
         RevealFullyPainted(
-            hwnd_,
-            SW_SHOW);
+            hwnd_);
     SetForegroundWindow(hwnd_);
 
     MSG msg{};